+++ docs/changelog.md
@@ -11,7 +11,8 @@
 - Scene renderables -- Image and Text
 - Scene cameras
 - Scene animators which process 'lerp' functions
-- Transform, collision and (simple) physics components
+- Transform, collision and physics components
+- A basic (swept AABB) physics system 
 - Sprite component and a sprite system, including animated sprites
 - Triggers that execute a callback function on collision enter, collide and exit
 - Buttons, with an option to provide a custom draw method